refactor pnp_diode benchmark to separate newton solver and mesh adaptivity and pnp coefficients to their own files for better organization.
